[fix] translations
