+++ connector_prestashop/unit/mapper.py
@@ -2,9 +2,9 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from openerp.addons.connector.unit.mapper import ExportMapper
+from openerp.addons.connector.unit.mapper import mapping
 
 
-# TODO: check if this stuff is needed for catalog export
 class PrestashopExportMapper(ExportMapper):
 
     def _map_direct(self, record, from_attr, to_attr):
@@ -21,17 +21,33 @@ class PrestashopExportMapper(ExportMapper):
 
 class TranslationPrestashopExportMapper(PrestashopExportMapper):
 
-    def convert(self, records_by_language, fields=None):
-        self.records_by_language = records_by_language
-        first_key = records_by_language.keys()[0]
-        self._convert(records_by_language[first_key], fields=fields)
-        self._data.update(self.convert_languages(self.translatable_fields))
+    @mapping
+    def translatable_fields(self, record):
+        fields = getattr(self, '_translatable_fields', [])
+        if fields:
+            translated_fields = self._convert_languages(
+                self._get_record_by_lang(record), fields)
+            return translated_fields
+        return {}
 
-    def convert_languages(self, records_by_language, translatable_fields):
+    def _get_record_by_lang(self, record):
+        # get the backend's languages
+        languages = self.backend_record.language_ids
+        records = {}
+        # for each languages:
+        for language in languages:
+            # get the translated record
+            record = record.with_context(
+                lang=language['code'])
+            # put it in the dict
+            records[language['prestashop_id']] = record
+        return records
+
+    def _convert_languages(self, records_by_language, translatable_fields):
         res = {}
         for from_attr, to_attr in translatable_fields:
             value = {'language': []}
-            for language_id, record in records_by_language.items():
+            for language_id, record in records_by_language.iteritems():
                 value['language'].append({
                     'attrs': {'id': str(language_id)},
                     'value': record[from_attr] or ''
