+++ connector_prestashop_catalog_manager/product_image.py
@@ -84,14 +84,14 @@ class ProductImageExport(PrestashopExporter):
     def _run(self, fields=None):
         """ Flow of the synchronization, implemented in inherited classes"""
         assert self.binding_id
-        assert self.erp_record
+        assert self.binding
 
         if self._has_to_skip():
             return
 
         # export the missing linked resources
         self._export_dependencies()
-        map_record = self.mapper.map_record(self.erp_record)
+        map_record = self.mapper.map_record(self.binding)
 
         if self.prestashop_id:
             record = map_record.values()
